Move some `apollo-composition` types into `apollo-federation-types`
