+++ Cargo.toml
@@ -1,3 +1,6 @@
 [workspace]
 members = ["apollo-federation-types", "apollo-composition"]
 resolver = "2"
+
+[workspace.dependencies]
+apollo-compiler = "1.0.0-beta.24"
