+++ apollo-composition/Cargo.toml
@@ -9,7 +9,7 @@ readme = "README.md"
 repository = "https://github.com/apollographql/federation-rs/"
 
 [dependencies]
-apollo-compiler = "1.0.0-beta.24"
+apollo-compiler = { workspace = true }
 apollo-federation = { version = "2.0.0-preview.3", git = "https://github.com/apollographql/router.git", branch = "next" }
-apollo-federation-types = { version = "0.15.0", path = "../apollo-federation-types" }
+apollo-federation-types = { version = "0.15.0", path = "../apollo-federation-types", features = ["composition"] }
 either = "1.12.0"
